Added changes on WelcomeWindow.java, where on create method the bottom navigation bar is set to visible.
Added changes the SecondFragment.java, on  create the bottom navigation bar is set to be invisible also overloaded the onCreate method, on onCreate method the onCreate option menu is set to true. set the search view on onCreate method on the top bar menu.

Added changes on learn.java, the first change was to set up listView and it has a controller class TabAdapter.java.