Designed overall look of the the matching game layout
The Text view on top is for game instruction, Followed by a layout intended for the matching English word with Xhosa word.

The Third layout contain textviews that contains Xhosa words that are to be used when choosing the word matching English word.

The submit button is to be used when the user is done matching the words.